High Level Design of an Online Food Ordering and Delivery System Uploaded.
Detailed High Level Design of an Online Food Ordering and Delivery System like Zomato and Swiggy Uploaded. In depth Final Design of this system is explained . Detailed Breakdown of each System as well as what are the communication protocols involved. Load Balancing, Database Selection ,Distributed Database ,Zookeeper etc., and Compression Algorithm chosen are explained in detail along with the reason of choosing them. Each Step of High Level Design is explained in detail.